feat: SEO 기초 세팅 추가 (#697)
- Root Layout에 Next.js Metadata API 기본 메타데이터 추가 (title, description, OG, Twitter Card)
- Home 페이지 canonical URL 메타데이터 추가
- Search 페이지 'use client' 분리 후 서버 컴포넌트에서 메타데이터 export
- robots.ts 생성 (크롤링 제어)
- sitemap.ts 생성 (공개 페이지: /, /search)
