+++ src/app/(home)/page.tsx
@@ -1,8 +1,13 @@
+import type { Metadata } from 'next';
 import Footer from '@/app/(home)/components/Footer/Footer';
 import HomeCarousel from '@/app/(home)/components/HomeCarousel/HomeCarousel';
 import LatestLessons from '@/app/(home)/components/LatestLessons/LatestLessons';
 import PopularGenre from '@/app/(home)/components/PopularGenre/PopularGenre';
 import UpcomingLessons from '@/app/(home)/components/UpcomingLessons/UpcomingLessons';
+
+export const metadata: Metadata = {
+  alternates: { canonical: '/' },
+};
 
 export default function Page() {
   return (
